# openapi/device_service.yaml
openapi: 3.0.3
info:
  title: Smart Device Service API
  description: API for executing user commands and smart home scenarios
  version: 1.0.0
servers:
  - url: https://warmhouse.com/api/v1
    description: Smart Device Service
paths:
  /commands:
    post:
      tags: [Commands]
      summary: Send command to device
      operationId: sendDeviceCommand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendCommandRequest'
      responses:
        '202':
          description: Command accepted for execution
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Device not found or unavailable
  /commands/{commandId}:
    get:
      tags: [Commands]
      summary: Get command status
      operationId: getCommandStatus
      parameters:
        - name: commandId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Command information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceCommand'
        '404':
          description: Command not found
...
components:
  schemas:
    SendCommandRequest:
      type: object
      required: [deviceId, command, sourceType]
      properties:
        deviceId:
          type: string
          format: uuid
        command:
          type: string
        parameters:
          type: object
          additionalProperties: true
        sourceType:
          type: string
          enum: [user, scenario, system]
        sourceId:
          type: string
          format: uuid
        priority:
          type: integer
          minimum: 0
          maximum: 10
          default: 0
...
    CommandResponse:
      type: object
      properties:
        commandId:
          type: string
          format: uuid
        status:
          type: string
          enum: [accepted, rejected]
        message:
          type: string
        timestamp:
          type: string
          format: date-time
    DeviceCommand:
      type: object
      properties:
        id:
          type: string
          format: uuid
        deviceId:
          type: string
          format: uuid
        sourceType:
          type: string
          enum: [user, scenario, system]
        sourceId:
          type: string
          format: uuid
        command:
          type: string
        parameters:
          type: object
        status:
          type: string
          enum: [sent, delivered, executed, failed, canceled]
        sentAt:
          type: string
          format: date-time
        executedAt:
          type: string
          format: date-time
        result:
          type: object
        priority:
          type: integer
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              message:
                type: string
              timestamp:
                type: string
                format: date-time
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []